api/requirements.txt
fastapi
uvicorn
requests
ruamel.yaml
docker